requirements.txt
# add requirements
mysql-connector-python 
